Use numext::sqrt in Householder.h.
This is to make it easier to apply to custom types - by using the
`numext` version, the user can specialize the function more easily.
Otherwise we require a `sqrt` function to be defined prior to including
the Eigen headers, which can be awkward and lead to header-include-order
issues.

Related to #2496