[SYCL] Fix some SYCL tests
Sigmoid failed in tensor_math because of the specializations in PacketMath.
The binary logic operators were casting floating types with rounding but they
are meant to do bitwise casting. The generic implementations of and, or, xor,
andnot are working as expected with SYCL.

tensor_builtin test could fail for certain seeds because of log
producing too small outputs for the test precision.

tensor_random could fail for certain seeds. The neighborhood check is
removed as it is not a safe way to check for the distribution.
This matches the behavior of the CUDA test.