Adjust overflow threshold bound for pow tests.
The original bound was too weak and did result in some overflows causing
our CI pipelines to fail (https://gitlab.com/libeigen/eigen_ci_cross_testing/-/pipelines/631839537).

In particular, it allowed `int32_t` `pow(256, 4)`, which overflowed to 0
for integers, and to an inexact result for float.  Here we adjust the
bounds.