Simpler range reduction strategy for atan<float>().
This saves a division in exchange for a couple of extra FMAs. The relative error is still <= 2 ulps, while speedup is 20-40% on x86. https://gitlab.com/libeigen/eigen/-/snippets/2421160

Also, this change refactors the inner polynomial approximations for atan<float>() and atan<double>() to separate fuinctions for future use in a more efficient implementation of atan2().