Fix serialization for non-compressed matrices.
The size of the data buffer was incorrect - it is not the number of
non-zeros, that is only true for compressed matrices.

It turned out that clang previously always converted the test matrices
to compressed mode because of the move-constructor, whereas this move
construction was optimized out by gcc.  Added an explicit move constructor
to fix this.